moved OpenStax::Exchange::Client functionality into OpenStax::Exchange
