--- lib/openstax/exchange/client.rb
@@ -1,8 +1,0 @@
-require "openstax/exchange/client/version"
-require "openstax/exchange/client/client"
-require "openstax/exchange/client/client_configuration"
-require "openstax/exchange/client/client_instance"
-require "openstax/exchange/client/real_client"
-require "openstax/exchange/client/fake_client"
-require "openstax/exchange/client/fake_client_configuration"
-require "openstax/exchange/client/exceptions"
