--- lib/openstax/exchange/client/client_configuration.rb
@@ -1,13 +1,0 @@
-module OpenStax
-  module Exchange
-    module Client
-      class ClientConfiguration
-        attr_accessor :platform_id
-        attr_accessor :platform_secret
-        attr_accessor :server_base_url
-        attr_accessor :server_port
-        attr_accessor :api_version
-      end
-    end
-  end
-end
