--- lib/openstax/exchange/client/client_instance.rb
@@ -1,17 +1,0 @@
-module OpenStax
-  module Exchange
-    module Client
-
-      module ClientInstance
-        def is_real_client?
-          raise NotImplementedError
-        end
-
-        def token
-          raise NotImplementedError
-        end
-      end
-
-    end
-  end
-end
