--- lib/openstax/exchange/client/version.rb
@@ -1,7 +1,0 @@
-module OpenStax
-  module Exchange
-    module Client
-      VERSION = "0.0.1"
-    end
-  end
-end
