+++ lib/openstax/exchange/client_configuration.rb
@@ -1,0 +1,13 @@
+module OpenStax
+  module Exchange
+
+    class ClientConfiguration
+      attr_accessor :platform_id
+      attr_accessor :platform_secret
+      attr_accessor :server_base_url
+      attr_accessor :server_port
+      attr_accessor :api_version
+    end
+
+  end
+end
