+++ lib/openstax/exchange/fake_client/fake_client_configuration.rb
@@ -1,14 +1,14 @@
 module OpenStax
   module Exchange
-    module Client
 
-      class FakeClientConfiguration
+    class FakeClient
+      class Configuration
         attr_accessor :registered_platforms
         attr_accessor :server_base_url
         attr_accessor :server_port
         attr_accessor :supported_api_versions
       end
-
     end
+
   end
 end
