+++ lib/openstax/exchange/version.rb
@@ -1,0 +1,5 @@
+module OpenStax
+  module Exchange
+    VERSION = "0.0.1"
+  end
+end
