--- spec/cassettes/OpenStax_Exchange_Client/internal_client_instance/_client_defaults_to_a_real_exchange_client.yml
@@ -1,48 +1,0 @@
----
-http_interactions:
-- request:
-    method: post
-    uri: http://123:abc@localhost:3003/oauth/token
-    body:
-      encoding: US-ASCII
-      string: grant_type=client_credentials
-    headers:
-      User-Agent:
-      - Faraday v0.9.1
-      Content-Type:
-      - application/x-www-form-urlencoded
-      Accept:
-      - ! '*/*'
-  response:
-    status:
-      code: 200
-      message: OK
-    headers:
-      X-Frame-Options:
-      - SAMEORIGIN
-      X-Xss-Protection:
-      - 1; mode=block
-      X-Content-Type-Options:
-      - nosniff
-      Cache-Control:
-      - no-store
-      Pragma:
-      - no-cache
-      Content-Type:
-      - application/json; charset=utf-8
-      Etag:
-      - W/"c9c086e39726c563b813fc5d9f34d5ad"
-      X-Request-Id:
-      - 176559b6-2092-40c1-8c87-895ecb278006
-      X-Runtime:
-      - '0.017105'
-      Connection:
-      - close
-      Server:
-      - thin
-    body:
-      encoding: US-ASCII
-      string: ! '{"access_token":"f626f66547a4290d9067c073ae811e42497bff0945c09c441922624565b18ffc","token_type":"bearer"}'
-    http_version: 
-  recorded_at: Mon, 12 Jan 2015 03:29:03 GMT
-recorded_with: VCR 2.9.3
