--- spec/cassettes/OpenStax_Exchange_Client/internal_client_instance/_reset_causes_a_new_client_object_to_be_returned_by_client.yml
@@ -1,48 +1,0 @@
----
-http_interactions:
-- request:
-    method: post
-    uri: http://123:abc@localhost:3003/oauth/token
-    body:
-      encoding: US-ASCII
-      string: grant_type=client_credentials
-    headers:
-      User-Agent:
-      - Faraday v0.9.1
-      Content-Type:
-      - application/x-www-form-urlencoded
-      Accept:
-      - ! '*/*'
-  response:
-    status:
-      code: 200
-      message: OK
-    headers:
-      X-Frame-Options:
-      - SAMEORIGIN
-      X-Xss-Protection:
-      - 1; mode=block
-      X-Content-Type-Options:
-      - nosniff
-      Cache-Control:
-      - no-store
-      Pragma:
-      - no-cache
-      Content-Type:
-      - application/json; charset=utf-8
-      Etag:
-      - W/"26e56695e30e7e6569a0235622a6fb0c"
-      X-Request-Id:
-      - 600193ef-ccea-4030-a173-8f4e39d97589
-      X-Runtime:
-      - '0.014589'
-      Connection:
-      - close
-      Server:
-      - thin
-    body:
-      encoding: US-ASCII
-      string: ! '{"access_token":"0cfa3c3244d6399c06158a4ed9f7d18fae725510f8b4ed414d22b5f2caa0a698","token_type":"bearer"}'
-    http_version: 
-  recorded_at: Mon, 12 Jan 2015 03:29:04 GMT
-recorded_with: VCR 2.9.3
