--- spec/cassettes/OpenStax_Exchange_Client_RealClient/behaves_like_exchange_client_api_v1/_create_identifier/success/creates_a_distinct_identifer_per_invokation.yml
@@ -1,146 +1,0 @@
----
-http_interactions:
-- request:
-    method: post
-    uri: http://123:abc@localhost:3003/oauth/token
-    body:
-      encoding: US-ASCII
-      string: grant_type=client_credentials
-    headers:
-      User-Agent:
-      - Faraday v0.9.1
-      Content-Type:
-      - application/x-www-form-urlencoded
-      Accept:
-      - ! '*/*'
-  response:
-    status:
-      code: 200
-      message: OK
-    headers:
-      X-Frame-Options:
-      - SAMEORIGIN
-      X-Xss-Protection:
-      - 1; mode=block
-      X-Content-Type-Options:
-      - nosniff
-      Cache-Control:
-      - no-store
-      Pragma:
-      - no-cache
-      Content-Type:
-      - application/json; charset=utf-8
-      Etag:
-      - W/"410e80860cc601540c682c69e1e3762f"
-      X-Request-Id:
-      - 45664369-0966-42aa-9467-27a8ead2a056
-      X-Runtime:
-      - '0.015172'
-      Connection:
-      - close
-      Server:
-      - thin
-    body:
-      encoding: US-ASCII
-      string: ! '{"access_token":"bc9b9db95898e10bdfec5e015642f69926e6df90a33a437b1216d0a4e8feffab","token_type":"bearer"}'
-    http_version: 
-  recorded_at: Mon, 12 Jan 2015 03:29:04 GMT
-- request:
-    method: post
-    uri: http://localhost:3003/api/identifiers
-    body:
-      encoding: US-ASCII
-      string: ''
-    headers:
-      User-Agent:
-      - Faraday v0.9.1
-      Accept:
-      - application/vnd.exchange.openstax.v1
-      Authorization:
-      - Bearer bc9b9db95898e10bdfec5e015642f69926e6df90a33a437b1216d0a4e8feffab
-      Content-Length:
-      - '0'
-  response:
-    status:
-      code: 201
-      message: Created
-    headers:
-      X-Frame-Options:
-      - SAMEORIGIN
-      X-Xss-Protection:
-      - 1; mode=block
-      X-Content-Type-Options:
-      - nosniff
-      Content-Type:
-      - application/json; charset=utf-8
-      Etag:
-      - W/"7a2374ba61cf288e9381149d1db77370"
-      Cache-Control:
-      - max-age=0, private, must-revalidate
-      Set-Cookie:
-      - _exercises_session=VWd0Y0hpMEZvTmdyQ2pSRE1SeDBvUlQxZXRpT3lZZW1LeUQzb2k2ZFFFTm9QUGhoM29NanZzcnliN2F6QlZpeVAwODRlWC9xVW5aM09sRmRXODVXUkE9PS0taDFoODM3amhCc2hDNUowSGVmaU10dz09--66d6a56912dd5fa8cec438dfb31151aa465fc0d5;
-        path=/; HttpOnly
-      - request_method=POST; path=/
-      X-Request-Id:
-      - fc98ff7e-f962-4cb8-ba5a-835795b349c0
-      X-Runtime:
-      - '0.124185'
-      Connection:
-      - close
-      Server:
-      - thin
-    body:
-      encoding: US-ASCII
-      string: ! '{"identifier":"d6a5d2cba08334cda0edefd6303108645999191e6b13de4abcfb93f9a801eb53"}'
-    http_version: 
-  recorded_at: Mon, 12 Jan 2015 03:29:04 GMT
-- request:
-    method: post
-    uri: http://localhost:3003/api/identifiers
-    body:
-      encoding: US-ASCII
-      string: ''
-    headers:
-      User-Agent:
-      - Faraday v0.9.1
-      Accept:
-      - application/vnd.exchange.openstax.v1
-      Authorization:
-      - Bearer bc9b9db95898e10bdfec5e015642f69926e6df90a33a437b1216d0a4e8feffab
-      Content-Length:
-      - '0'
-  response:
-    status:
-      code: 201
-      message: Created
-    headers:
-      X-Frame-Options:
-      - SAMEORIGIN
-      X-Xss-Protection:
-      - 1; mode=block
-      X-Content-Type-Options:
-      - nosniff
-      Content-Type:
-      - application/json; charset=utf-8
-      Etag:
-      - W/"9b7b118f6951f14caaeb576b0126c94c"
-      Cache-Control:
-      - max-age=0, private, must-revalidate
-      Set-Cookie:
-      - _exercises_session=YzVhaDRDOVhIRGpaenJFc29zL0EwRjhGZGIvVEZ4dnJOTWlTRXBRSjZpT1AwNXQ0V2psUjdsNzVaaDc1YXlYQlNQR2F3MHRuYWxGTUwwL2JMcnY3Unc9PS0tbTI3SVpYcW11UnBmZ0JHQWNvejBaZz09--b604e1d19eb743218004e93c0f53218b4aad366e;
-        path=/; HttpOnly
-      - request_method=POST; path=/
-      X-Request-Id:
-      - fff6a01a-aa5d-4d40-b233-76f8ab28963f
-      X-Runtime:
-      - '0.043432'
-      Connection:
-      - close
-      Server:
-      - thin
-    body:
-      encoding: US-ASCII
-      string: ! '{"identifier":"d31491426e8b310f55027042a95be2f4d9f9b4b321fa26237668c2fc01aedb18"}'
-    http_version: 
-  recorded_at: Mon, 12 Jan 2015 03:29:04 GMT
-recorded_with: VCR 2.9.3
