--- spec/cassettes/OpenStax_Exchange_Client_RealClient/behaves_like_exchange_client_api_v1/_create_identifier/success/creates_and_returns_a_new_identifier.yml
@@ -1,97 +1,0 @@
----
-http_interactions:
-- request:
-    method: post
-    uri: http://123:abc@localhost:3003/oauth/token
-    body:
-      encoding: US-ASCII
-      string: grant_type=client_credentials
-    headers:
-      User-Agent:
-      - Faraday v0.9.1
-      Content-Type:
-      - application/x-www-form-urlencoded
-      Accept:
-      - ! '*/*'
-  response:
-    status:
-      code: 200
-      message: OK
-    headers:
-      X-Frame-Options:
-      - SAMEORIGIN
-      X-Xss-Protection:
-      - 1; mode=block
-      X-Content-Type-Options:
-      - nosniff
-      Cache-Control:
-      - no-store
-      Pragma:
-      - no-cache
-      Content-Type:
-      - application/json; charset=utf-8
-      Etag:
-      - W/"6d4c6cee2f26f6aaf5ad5fcd53baee62"
-      X-Request-Id:
-      - d2415e21-d66a-44df-81cb-73ec323597c9
-      X-Runtime:
-      - '0.020114'
-      Connection:
-      - close
-      Server:
-      - thin
-    body:
-      encoding: US-ASCII
-      string: ! '{"access_token":"690c3f22c3dbdbb0ce5544441e8954b98ca3d643770997fca5bd7ad74fdb2bb3","token_type":"bearer"}'
-    http_version: 
-  recorded_at: Mon, 12 Jan 2015 03:29:04 GMT
-- request:
-    method: post
-    uri: http://localhost:3003/api/identifiers
-    body:
-      encoding: US-ASCII
-      string: ''
-    headers:
-      User-Agent:
-      - Faraday v0.9.1
-      Accept:
-      - application/vnd.exchange.openstax.v1
-      Authorization:
-      - Bearer 690c3f22c3dbdbb0ce5544441e8954b98ca3d643770997fca5bd7ad74fdb2bb3
-      Content-Length:
-      - '0'
-  response:
-    status:
-      code: 201
-      message: Created
-    headers:
-      X-Frame-Options:
-      - SAMEORIGIN
-      X-Xss-Protection:
-      - 1; mode=block
-      X-Content-Type-Options:
-      - nosniff
-      Content-Type:
-      - application/json; charset=utf-8
-      Etag:
-      - W/"4614e9310f65e35729585ba1a30b90ce"
-      Cache-Control:
-      - max-age=0, private, must-revalidate
-      Set-Cookie:
-      - _exercises_session=a2o3Q1lVVkxqN3kxWXVGeEtqZzJmejJsNGF6NFlkbmJHMS9VNGNCUVFTbmxxdC91R25pbEJvQUhEUXVXK1ZvWDBraE1qOUtBa3JleGU5STZFZTVlTFE9PS0tUVNTVGVKSGl6bGF6cGhGK2ZYV0g5UT09--00b1e652b9079eac926fdbe792c026693b94d87f;
-        path=/; HttpOnly
-      - request_method=POST; path=/
-      X-Request-Id:
-      - d6375424-0ed0-4694-9fb6-628ac3ee4804
-      X-Runtime:
-      - '0.040582'
-      Connection:
-      - close
-      Server:
-      - thin
-    body:
-      encoding: US-ASCII
-      string: ! '{"identifier":"71ca83f8dd5e7f3721c60e28237cb06815e20231676f498ef595e2a857c7c6f8"}'
-    http_version: 
-  recorded_at: Mon, 12 Jan 2015 03:29:04 GMT
-recorded_with: VCR 2.9.3
