--- spec/cassettes/OpenStax_Exchange_Client_RealClient/behaves_like_exchange_client_api_v1/_create_multiple_choice/duplicate_identifer_resource_trial_triplet/raises_an_exception.yml
@@ -1,194 +1,0 @@
----
-http_interactions:
-- request:
-    method: post
-    uri: http://123:abc@localhost:3003/oauth/token
-    body:
-      encoding: US-ASCII
-      string: grant_type=client_credentials
-    headers:
-      User-Agent:
-      - Faraday v0.9.1
-      Content-Type:
-      - application/x-www-form-urlencoded
-      Accept:
-      - ! '*/*'
-  response:
-    status:
-      code: 200
-      message: OK
-    headers:
-      X-Frame-Options:
-      - SAMEORIGIN
-      X-Xss-Protection:
-      - 1; mode=block
-      X-Content-Type-Options:
-      - nosniff
-      Cache-Control:
-      - no-store
-      Pragma:
-      - no-cache
-      Content-Type:
-      - application/json; charset=utf-8
-      Etag:
-      - W/"9a8b3003cf7aacb6d0484841a41f5896"
-      X-Request-Id:
-      - 881c9ea6-390b-415d-a21c-ad7305a29f61
-      X-Runtime:
-      - '0.015520'
-      Connection:
-      - close
-      Server:
-      - thin
-    body:
-      encoding: US-ASCII
-      string: ! '{"access_token":"36e312a3b772c95afe3d5df1e718964f7ef398de247e8d3d3abc7f52215888c2","token_type":"bearer"}'
-    http_version: 
-  recorded_at: Mon, 12 Jan 2015 03:29:05 GMT
-- request:
-    method: post
-    uri: http://localhost:3003/api/identifiers
-    body:
-      encoding: US-ASCII
-      string: ''
-    headers:
-      User-Agent:
-      - Faraday v0.9.1
-      Accept:
-      - application/vnd.exchange.openstax.v1
-      Authorization:
-      - Bearer 36e312a3b772c95afe3d5df1e718964f7ef398de247e8d3d3abc7f52215888c2
-      Content-Length:
-      - '0'
-  response:
-    status:
-      code: 201
-      message: Created
-    headers:
-      X-Frame-Options:
-      - SAMEORIGIN
-      X-Xss-Protection:
-      - 1; mode=block
-      X-Content-Type-Options:
-      - nosniff
-      Content-Type:
-      - application/json; charset=utf-8
-      Etag:
-      - W/"7f191d67306cb64285e4d90532a8e835"
-      Cache-Control:
-      - max-age=0, private, must-revalidate
-      Set-Cookie:
-      - _exercises_session=REVmRmFrbDNZZDFydk82UURyZEVYTGJSQ1g3VTU2cnZJSU4xbmFyN2lFZFNKM3RNekJHbnl2eWZJT2dYWUZnVnpnd2hlM0FYUnZyNDRKc3B2N1R5a1E9PS0tR0hhWE10ZENmZWMrYUlFdEF6V0Rudz09--903dff2f435ec633747cf06b13f8a2806a156ce9;
-        path=/; HttpOnly
-      - request_method=POST; path=/
-      X-Request-Id:
-      - f93614be-60f9-4e66-8f6c-2901b10b8daf
-      X-Runtime:
-      - '0.052542'
-      Connection:
-      - close
-      Server:
-      - thin
-    body:
-      encoding: US-ASCII
-      string: ! '{"identifier":"e1856bcca610558a72e337d8743a2ac4ac798a20894b4a0450046183ee8fb4fb"}'
-    http_version: 
-  recorded_at: Mon, 12 Jan 2015 03:29:05 GMT
-- request:
-    method: post
-    uri: http://localhost:3003/api/events/platforms/multiple_choices
-    body:
-      encoding: UTF-8
-      string: ! '{"identifier":"e1856bcca610558a72e337d8743a2ac4ac798a20894b4a0450046183ee8fb4fb","resource":"http://exercises.openstax.org/exercises/1234","trial":"1","answer":"answer_string"}'
-    headers:
-      User-Agent:
-      - Faraday v0.9.1
-      Accept:
-      - application/vnd.exchange.openstax.v1
-      Authorization:
-      - Bearer 36e312a3b772c95afe3d5df1e718964f7ef398de247e8d3d3abc7f52215888c2
-      Content-Type:
-      - application/x-www-form-urlencoded
-  response:
-    status:
-      code: 201
-      message: Created
-    headers:
-      X-Frame-Options:
-      - SAMEORIGIN
-      X-Xss-Protection:
-      - 1; mode=block
-      X-Content-Type-Options:
-      - nosniff
-      Content-Type:
-      - application/json; charset=utf-8
-      Etag:
-      - W/"d9aae3fd8a8ac0b06d0db783fc85de8e"
-      Cache-Control:
-      - max-age=0, private, must-revalidate
-      Set-Cookie:
-      - _exercises_session=R2NUdWhlaDFUUWw3THQ3bkpSM0RzY1YrS3pyeThBbTlxQldwQXNLU1FNY2V3bC9kMXAwa2NIQVdZNWpPWmduUy9pOWlsRzdaK25RamZ2U1MyekhKRXc9PS0tUFQvdlpOelNWLzRxdmxteC9QQ3lzUT09--c6e538e006b02578e82593f434520099cb8cf0ec;
-        path=/; HttpOnly
-      - request_method=POST; path=/
-      X-Request-Id:
-      - 2a8b7c19-626d-40f3-bcb1-983b0c1b3682
-      X-Runtime:
-      - '0.077004'
-      Connection:
-      - close
-      Server:
-      - thin
-    body:
-      encoding: US-ASCII
-      string: ! '{"identifier":"e1856bcca610558a72e337d8743a2ac4ac798a20894b4a0450046183ee8fb4fb","resource":"http://exercises.openstax.org/exercises/1234","trial":"1","created_at":"2015-01-12T03:29:05.623Z","answer_type":"multiple-choice","answer":"answer_string"}'
-    http_version: 
-  recorded_at: Mon, 12 Jan 2015 03:29:05 GMT
-- request:
-    method: post
-    uri: http://localhost:3003/api/events/platforms/multiple_choices
-    body:
-      encoding: UTF-8
-      string: ! '{"identifier":"e1856bcca610558a72e337d8743a2ac4ac798a20894b4a0450046183ee8fb4fb","resource":"http://exercises.openstax.org/exercises/1234","trial":"1","answer":"answer_string"}'
-    headers:
-      User-Agent:
-      - Faraday v0.9.1
-      Accept:
-      - application/vnd.exchange.openstax.v1
-      Authorization:
-      - Bearer 36e312a3b772c95afe3d5df1e718964f7ef398de247e8d3d3abc7f52215888c2
-      Content-Type:
-      - application/x-www-form-urlencoded
-  response:
-    status:
-      code: 422
-      message: Unprocessable Entity
-    headers:
-      X-Frame-Options:
-      - SAMEORIGIN
-      X-Xss-Protection:
-      - 1; mode=block
-      X-Content-Type-Options:
-      - nosniff
-      Content-Type:
-      - application/json; charset=utf-8
-      Cache-Control:
-      - no-cache
-      Set-Cookie:
-      - _exercises_session=U0tpa0lqSkx3Y3RPcFVhL24yY0dNcThIQ3ZhK0dxbURzbWJ6ZlhSdG1WK1o3VTdnRlloVlZZVHN1dEtCYlIxanlyd3FFVUx3OVdhd2NkUFg2dldGaWc9PS0taDJtTDN6LzZTcnB4bjV3aDdETmlmUT09--205f72ab300e81efc120afb4696c6e0f33c64c1c;
-        path=/; HttpOnly
-      - request_method=POST; path=/
-      X-Request-Id:
-      - 23f7fe8b-fe31-4c57-a410-72a5b1defc79
-      X-Runtime:
-      - '0.039955'
-      Connection:
-      - close
-      Server:
-      - thin
-    body:
-      encoding: US-ASCII
-      string: ! '[{"code":"taken","data":{"model":{"id":null,"identifier_id":864,"resource_id":5,"trial":"1","due_at":null,"created_at":null,"updated_at":null},"attribute":"trial"},"kind":"activerecord","message":"has
-        already been taken","offending_inputs":"trial"}]'
-    http_version: 
-  recorded_at: Mon, 12 Jan 2015 03:29:05 GMT
-recorded_with: VCR 2.9.3
