--- spec/cassettes/OpenStax_Exchange_Client_RealClient/behaves_like_exchange_client_api_v1/_create_multiple_choice/invalid_resource_string/raises_an_exception.yml
@@ -1,145 +1,0 @@
----
-http_interactions:
-- request:
-    method: post
-    uri: http://123:abc@localhost:3003/oauth/token
-    body:
-      encoding: US-ASCII
-      string: grant_type=client_credentials
-    headers:
-      User-Agent:
-      - Faraday v0.9.1
-      Content-Type:
-      - application/x-www-form-urlencoded
-      Accept:
-      - ! '*/*'
-  response:
-    status:
-      code: 200
-      message: OK
-    headers:
-      X-Frame-Options:
-      - SAMEORIGIN
-      X-Xss-Protection:
-      - 1; mode=block
-      X-Content-Type-Options:
-      - nosniff
-      Cache-Control:
-      - no-store
-      Pragma:
-      - no-cache
-      Content-Type:
-      - application/json; charset=utf-8
-      Etag:
-      - W/"bf30c872593d64cf2d554a65ebf08a2b"
-      X-Request-Id:
-      - ac2ce795-e10f-4390-b3a2-987d42df049b
-      X-Runtime:
-      - '0.053544'
-      Connection:
-      - close
-      Server:
-      - thin
-    body:
-      encoding: US-ASCII
-      string: ! '{"access_token":"903a38d016bcea827d121bab4c92fb22afab675e25d123f6c6014534ce3f7f7e","token_type":"bearer"}'
-    http_version: 
-  recorded_at: Mon, 12 Jan 2015 03:29:05 GMT
-- request:
-    method: post
-    uri: http://localhost:3003/api/identifiers
-    body:
-      encoding: US-ASCII
-      string: ''
-    headers:
-      User-Agent:
-      - Faraday v0.9.1
-      Accept:
-      - application/vnd.exchange.openstax.v1
-      Authorization:
-      - Bearer 903a38d016bcea827d121bab4c92fb22afab675e25d123f6c6014534ce3f7f7e
-      Content-Length:
-      - '0'
-  response:
-    status:
-      code: 201
-      message: Created
-    headers:
-      X-Frame-Options:
-      - SAMEORIGIN
-      X-Xss-Protection:
-      - 1; mode=block
-      X-Content-Type-Options:
-      - nosniff
-      Content-Type:
-      - application/json; charset=utf-8
-      Etag:
-      - W/"bda9d0d9930bda1729884601990fe356"
-      Cache-Control:
-      - max-age=0, private, must-revalidate
-      Set-Cookie:
-      - _exercises_session=dGEwN0hqNy9DMmV0VzBLVVV6ZkF6ek53WmE3M1ZyRFM5QlZ1ZjJPYzByaHNuMzhxNlBFNGVUTHMvem1zS1AvSVN6YURRSU80TzByRndqMUJGSFk0YXc9PS0tekNDejRqY2ZOSEVqTFNWRjh2bTdZdz09--3cd6c67781fb6d43d4487cbf3d3241af1f98e3a3;
-        path=/; HttpOnly
-      - request_method=POST; path=/
-      X-Request-Id:
-      - 09db497c-2c64-4b75-a6e6-0299275ed43a
-      X-Runtime:
-      - '0.041985'
-      Connection:
-      - close
-      Server:
-      - thin
-    body:
-      encoding: US-ASCII
-      string: ! '{"identifier":"70eb97aa529513d2686a75d613b4d3795965f0c0bb11e272b929a965f1adf10f"}'
-    http_version: 
-  recorded_at: Mon, 12 Jan 2015 03:29:05 GMT
-- request:
-    method: post
-    uri: http://localhost:3003/api/events/platforms/multiple_choices
-    body:
-      encoding: UTF-8
-      string: ! '{"identifier":"70eb97aa529513d2686a75d613b4d3795965f0c0bb11e272b929a965f1adf10f","resource":"http://example.com/exercises/1234","trial":"1","answer":"answer_string"}'
-    headers:
-      User-Agent:
-      - Faraday v0.9.1
-      Accept:
-      - application/vnd.exchange.openstax.v1
-      Authorization:
-      - Bearer 903a38d016bcea827d121bab4c92fb22afab675e25d123f6c6014534ce3f7f7e
-      Content-Type:
-      - application/x-www-form-urlencoded
-  response:
-    status:
-      code: 422
-      message: Unprocessable Entity
-    headers:
-      X-Frame-Options:
-      - SAMEORIGIN
-      X-Xss-Protection:
-      - 1; mode=block
-      X-Content-Type-Options:
-      - nosniff
-      Content-Type:
-      - application/json; charset=utf-8
-      Cache-Control:
-      - no-cache
-      Set-Cookie:
-      - _exercises_session=cEYyVDEyMW1CRVZuVVp5NTFjVTRUWDNlOWJ2T0o5RHVHZ0hMOHp1alFmb3piNTFuNzc4bzlaV2oyanYzR2x6NDdwV1VnSkpsRW5rVnl6WUIwSWM4MXc9PS0tQnAzYlgyN2pjVzZlcUZyUGZWb3YzZz09--64b62df905620220f874b4fe60ed267d9123713d;
-        path=/; HttpOnly
-      - request_method=POST; path=/
-      X-Request-Id:
-      - 8537c8cc-7438-4174-a52b-40bc5ab90de6
-      X-Runtime:
-      - '0.036261'
-      Connection:
-      - close
-      Server:
-      - thin
-    body:
-      encoding: US-ASCII
-      string: ! '[{"code":"blank","data":{"model":{"id":null,"identifier_id":865,"resource_id":null,"trial":"1","due_at":null,"created_at":null,"updated_at":null},"attribute":"resource"},"kind":"activerecord","message":"can''t
-        be blank","offending_inputs":"resource"}]'
-    http_version: 
-  recorded_at: Mon, 12 Jan 2015 03:29:05 GMT
-recorded_with: VCR 2.9.3
