--- spec/cassettes/OpenStax_Exchange_Client_RealClient/behaves_like_exchange_client_api_v1/_create_multiple_choice/success/creates_a_multiple_choice_response_associated_with_the_given_identifier.yml
@@ -1,146 +1,0 @@
----
-http_interactions:
-- request:
-    method: post
-    uri: http://123:abc@localhost:3003/oauth/token
-    body:
-      encoding: US-ASCII
-      string: grant_type=client_credentials
-    headers:
-      User-Agent:
-      - Faraday v0.9.1
-      Content-Type:
-      - application/x-www-form-urlencoded
-      Accept:
-      - ! '*/*'
-  response:
-    status:
-      code: 200
-      message: OK
-    headers:
-      X-Frame-Options:
-      - SAMEORIGIN
-      X-Xss-Protection:
-      - 1; mode=block
-      X-Content-Type-Options:
-      - nosniff
-      Cache-Control:
-      - no-store
-      Pragma:
-      - no-cache
-      Content-Type:
-      - application/json; charset=utf-8
-      Etag:
-      - W/"7e3c21201d1b2814c3851263ecc3f1a8"
-      X-Request-Id:
-      - bcd4582e-b2ac-41c7-a952-92da494e18dd
-      X-Runtime:
-      - '0.015486'
-      Connection:
-      - close
-      Server:
-      - thin
-    body:
-      encoding: US-ASCII
-      string: ! '{"access_token":"09aeb93ed921bb6cf5cfadd330454487e3608480d5f407f40188b4d8bf75b4c8","token_type":"bearer"}'
-    http_version: 
-  recorded_at: Mon, 12 Jan 2015 03:29:04 GMT
-- request:
-    method: post
-    uri: http://localhost:3003/api/identifiers
-    body:
-      encoding: US-ASCII
-      string: ''
-    headers:
-      User-Agent:
-      - Faraday v0.9.1
-      Accept:
-      - application/vnd.exchange.openstax.v1
-      Authorization:
-      - Bearer 09aeb93ed921bb6cf5cfadd330454487e3608480d5f407f40188b4d8bf75b4c8
-      Content-Length:
-      - '0'
-  response:
-    status:
-      code: 201
-      message: Created
-    headers:
-      X-Frame-Options:
-      - SAMEORIGIN
-      X-Xss-Protection:
-      - 1; mode=block
-      X-Content-Type-Options:
-      - nosniff
-      Content-Type:
-      - application/json; charset=utf-8
-      Etag:
-      - W/"f47b695aa3fe46ce8393b89ad46088a6"
-      Cache-Control:
-      - max-age=0, private, must-revalidate
-      Set-Cookie:
-      - _exercises_session=YmNGR3NGZ2ZGbXIxclk0VERWK0c1UCtPUVdWOW5LWnRHcVRzTEdSRWZLMC81V0ZRWXFlZDVrb2dQTXlNM25wWnYrVjMxY2RJWEJLL0pkN0IxcnB0eGc9PS0tRW96blZYVEloSy9jZVQva0QzdGFydz09--40fc3d839fc40189ad09fcb10eff966c8976c1a4;
-        path=/; HttpOnly
-      - request_method=POST; path=/
-      X-Request-Id:
-      - e3164cb3-00d8-47b2-b5b4-f9a8a8a3aca3
-      X-Runtime:
-      - '0.042579'
-      Connection:
-      - close
-      Server:
-      - thin
-    body:
-      encoding: US-ASCII
-      string: ! '{"identifier":"4ef0a654fbb01f5e3ffec5a1fe41b60304aa36d899db17c76efdc60d3dd5e453"}'
-    http_version: 
-  recorded_at: Mon, 12 Jan 2015 03:29:04 GMT
-- request:
-    method: post
-    uri: http://localhost:3003/api/events/platforms/multiple_choices
-    body:
-      encoding: UTF-8
-      string: ! '{"identifier":"4ef0a654fbb01f5e3ffec5a1fe41b60304aa36d899db17c76efdc60d3dd5e453","resource":"http://exercises.openstax.org/exercises/1234","trial":"1","answer":"answer_string"}'
-    headers:
-      User-Agent:
-      - Faraday v0.9.1
-      Accept:
-      - application/vnd.exchange.openstax.v1
-      Authorization:
-      - Bearer 09aeb93ed921bb6cf5cfadd330454487e3608480d5f407f40188b4d8bf75b4c8
-      Content-Type:
-      - application/x-www-form-urlencoded
-  response:
-    status:
-      code: 201
-      message: Created
-    headers:
-      X-Frame-Options:
-      - SAMEORIGIN
-      X-Xss-Protection:
-      - 1; mode=block
-      X-Content-Type-Options:
-      - nosniff
-      Content-Type:
-      - application/json; charset=utf-8
-      Etag:
-      - W/"901f7b5ecc6b6cf2b7fa83038c96e00e"
-      Cache-Control:
-      - max-age=0, private, must-revalidate
-      Set-Cookie:
-      - _exercises_session=aXdrOUhRM0pFRDFGK01lUDI0RW1QdS9wSS9lMDRkcVdUbGhuWUh1czFEenN5R2FHMzRhUmk2SmdubXBuL2k4c0JXbG1WTUZsNCswTTlqRWQwK1hMcnc9PS0tK1hjNXBRTzlEVHFUNXN0WUlRdzdVQT09--f103ed9fa568f71585a9c6b5f9471108b912042e;
-        path=/; HttpOnly
-      - request_method=POST; path=/
-      X-Request-Id:
-      - 5134106c-2c81-4a3b-ab40-739c1f061ee3
-      X-Runtime:
-      - '0.081367'
-      Connection:
-      - close
-      Server:
-      - thin
-    body:
-      encoding: US-ASCII
-      string: ! '{"identifier":"4ef0a654fbb01f5e3ffec5a1fe41b60304aa36d899db17c76efdc60d3dd5e453","resource":"http://exercises.openstax.org/exercises/1234","trial":"1","created_at":"2015-01-12T03:29:04.545Z","answer_type":"multiple-choice","answer":"answer_string"}'
-    http_version: 
-  recorded_at: Mon, 12 Jan 2015 03:29:04 GMT
-recorded_with: VCR 2.9.3
