--- spec/cassettes/OpenStax_Exchange_Client_RealClient/behaves_like_exchange_client_api_v1/_initialize/invalid_platform_id/raises_an_exception.yml
@@ -1,52 +1,0 @@
----
-http_interactions:
-- request:
-    method: post
-    uri: http://999:abc@localhost:3003/oauth/token
-    body:
-      encoding: US-ASCII
-      string: grant_type=client_credentials
-    headers:
-      User-Agent:
-      - Faraday v0.9.1
-      Content-Type:
-      - application/x-www-form-urlencoded
-      Accept:
-      - ! '*/*'
-  response:
-    status:
-      code: 401
-      message: Unauthorized
-    headers:
-      X-Frame-Options:
-      - SAMEORIGIN
-      X-Xss-Protection:
-      - 1; mode=block
-      X-Content-Type-Options:
-      - nosniff
-      Cache-Control:
-      - no-store
-      Pragma:
-      - no-cache
-      Content-Type:
-      - application/json; charset=utf-8
-      Www-Authenticate:
-      - Bearer realm="Doorkeeper", error="invalid_client", error_description="Client
-        authentication failed due to unknown client, no client authentication included,
-        or unsupported authentication method."
-      X-Request-Id:
-      - d7dbfcf7-efe1-48bb-841e-36c6c62b659f
-      X-Runtime:
-      - '0.010380'
-      Connection:
-      - close
-      Server:
-      - thin
-    body:
-      encoding: US-ASCII
-      string: ! '{"error":"invalid_client","error_description":"Client authentication
-        failed due to unknown client, no client authentication included, or unsupported
-        authentication method."}'
-    http_version: 
-  recorded_at: Mon, 12 Jan 2015 03:29:04 GMT
-recorded_with: VCR 2.9.3
