--- spec/cassettes/OpenStax_Exchange_Client_RealClient/behaves_like_exchange_client_api_v1/_initialize/invalid_platform_secret/raises_an_exception.yml
@@ -1,52 +1,0 @@
----
-http_interactions:
-- request:
-    method: post
-    uri: http://123:not_the_secret@localhost:3003/oauth/token
-    body:
-      encoding: US-ASCII
-      string: grant_type=client_credentials
-    headers:
-      User-Agent:
-      - Faraday v0.9.1
-      Content-Type:
-      - application/x-www-form-urlencoded
-      Accept:
-      - ! '*/*'
-  response:
-    status:
-      code: 401
-      message: Unauthorized
-    headers:
-      X-Frame-Options:
-      - SAMEORIGIN
-      X-Xss-Protection:
-      - 1; mode=block
-      X-Content-Type-Options:
-      - nosniff
-      Cache-Control:
-      - no-store
-      Pragma:
-      - no-cache
-      Content-Type:
-      - application/json; charset=utf-8
-      Www-Authenticate:
-      - Bearer realm="Doorkeeper", error="invalid_client", error_description="Client
-        authentication failed due to unknown client, no client authentication included,
-        or unsupported authentication method."
-      X-Request-Id:
-      - e4dfd38c-25d1-457a-9f4c-04de186dad6d
-      X-Runtime:
-      - '0.010286'
-      Connection:
-      - close
-      Server:
-      - thin
-    body:
-      encoding: US-ASCII
-      string: ! '{"error":"invalid_client","error_description":"Client authentication
-        failed due to unknown client, no client authentication included, or unsupported
-        authentication method."}'
-    http_version: 
-  recorded_at: Mon, 12 Jan 2015 03:29:04 GMT
-recorded_with: VCR 2.9.3
