--- spec/cassettes/OpenStax_Exchange_Client_RealClient/behaves_like_exchange_client_api_v1/_initialize/success/initializes_the_authentication_token.yml
@@ -1,48 +1,0 @@
----
-http_interactions:
-- request:
-    method: post
-    uri: http://123:abc@localhost:3003/oauth/token
-    body:
-      encoding: US-ASCII
-      string: grant_type=client_credentials
-    headers:
-      User-Agent:
-      - Faraday v0.9.1
-      Content-Type:
-      - application/x-www-form-urlencoded
-      Accept:
-      - ! '*/*'
-  response:
-    status:
-      code: 200
-      message: OK
-    headers:
-      X-Frame-Options:
-      - SAMEORIGIN
-      X-Xss-Protection:
-      - 1; mode=block
-      X-Content-Type-Options:
-      - nosniff
-      Cache-Control:
-      - no-store
-      Pragma:
-      - no-cache
-      Content-Type:
-      - application/json; charset=utf-8
-      Etag:
-      - W/"124357e8d6a0903269c6cb4f4c5aacb8"
-      X-Request-Id:
-      - 6becf4b6-c6e9-4e4d-aadb-9f65b6f88024
-      X-Runtime:
-      - '0.014747'
-      Connection:
-      - close
-      Server:
-      - thin
-    body:
-      encoding: US-ASCII
-      string: ! '{"access_token":"4dfeede6d1933847e6ce7216af36ad638b2cb036906b41a2ad38eaa51ed0d23e","token_type":"bearer"}'
-    http_version: 
-  recorded_at: Mon, 12 Jan 2015 03:29:04 GMT
-recorded_with: VCR 2.9.3
