+++ spec/lib/openstax/exchange/client/client_configuration_spec.rb
@@ -1,9 +1,9 @@
 require 'spec_helper'
 
-describe OpenStax::Exchange::Client, vcr: VCR_OPTS do
+describe OpenStax::Exchange, vcr: VCR_OPTS do
 
   before(:each) do
-    OpenStax::Exchange::Client.configure do |config|
+    OpenStax::Exchange.configure do |config|
       config.platform_id     = DEFAULT_CLIENT_PLATFORM_ID
       config.platform_secret = DEFAULT_CLIENT_PLATFORM_SECRET
       config.server_base_url = DEFAULT_CLIENT_SERVER_BASE_URL
@@ -11,7 +11,7 @@ describe OpenStax::Exchange::Client, vcr: VCR_OPTS do
       config.api_version     = DEFAULT_CLIENT_API_VERSION
     end
 
-    OpenStax::Exchange::Client::FakeClient.configure do |config|
+    OpenStax::Exchange::FakeClient.configure do |config|
       config.registered_platforms   = {DEFAULT_CLIENT_PLATFORM_ID => DEFAULT_CLIENT_PLATFORM_SECRET}
       config.server_base_url        = DEFAULT_CLIENT_SERVER_BASE_URL
       config.server_port            = DEFAULT_CLIENT_SERVER_PORT
@@ -21,38 +21,38 @@ describe OpenStax::Exchange::Client, vcr: VCR_OPTS do
 
   context "internal client instance" do
     it "#client defaults to a real exchange client" do
-      client = OpenStax::Exchange::Client.send(:client)
+      client = OpenStax::Exchange.send(:client)
       expect(client.is_real_client?).to be_truthy
     end
 
     it "#client returns the same client object on successive calls" do
-      client_object_id1 = OpenStax::Exchange::Client.send(:client).object_id
-      client_object_id2 = OpenStax::Exchange::Client.send(:client).object_id
+      client_object_id1 = OpenStax::Exchange.send(:client).object_id
+      client_object_id2 = OpenStax::Exchange.send(:client).object_id
       expect(client_object_id1).to eq(client_object_id2)
     end
 
     it "#reset! causes a new client object to be returned by #client" do
-      client_object_id1 = OpenStax::Exchange::Client.send(:client).object_id
-      OpenStax::Exchange::Client.reset!
-      client_object_id2 = OpenStax::Exchange::Client.send(:client).object_id
+      client_object_id1 = OpenStax::Exchange.send(:client).object_id
+      OpenStax::Exchange.reset!
+      client_object_id2 = OpenStax::Exchange.send(:client).object_id
       expect(client_object_id1).to_not eq(client_object_id2)
     end
   end
 
   context "client instance configuration" do
     before(:each) do
-      OpenStax::Exchange::Client.reset!
+      OpenStax::Exchange.reset!
     end
 
     it "can be configured to use a real exchange client" do
-      OpenStax::Exchange::Client.use_real_client
-      client = OpenStax::Exchange::Client.send(:client)
+      OpenStax::Exchange.use_real_client
+      client = OpenStax::Exchange.send(:client)
       expect(client.is_real_client?).to be_truthy
     end
 
     it "can be configured to use a fake exchange client" do
-      OpenStax::Exchange::Client.use_fake_client
-      client = OpenStax::Exchange::Client.send(:client)
+      OpenStax::Exchange.use_fake_client
+      client = OpenStax::Exchange.send(:client)
       expect(client.is_real_client?).to be_falsy
     end
   end
