+++ spec/lib/openstax/exchange/client/fake_client_v1_spec.rb
@@ -1,9 +1,9 @@
 require 'spec_helper'
 
-describe OpenStax::Exchange::Client::FakeClient do
+describe OpenStax::Exchange::FakeClient do
   before(:each) do
-    OpenStax::Exchange::Client.use_fake_client
-    OpenStax::Exchange::Client::FakeClient.configure do |config|
+    OpenStax::Exchange.use_fake_client
+    OpenStax::Exchange::FakeClient.configure do |config|
       config.registered_platforms   = {DEFAULT_CLIENT_PLATFORM_ID => DEFAULT_CLIENT_PLATFORM_SECRET}
       config.server_base_url        = DEFAULT_CLIENT_SERVER_BASE_URL
       config.server_port            = DEFAULT_CLIENT_SERVER_PORT
