+++ spec/lib/openstax/exchange/client/real_client_v1_spec.rb
@@ -1,8 +1,8 @@
 require 'spec_helper'
 
-describe OpenStax::Exchange::Client::RealClient, vcr: VCR_OPTS do
+describe OpenStax::Exchange::RealClient, vcr: VCR_OPTS do
   before(:each) do
-    OpenStax::Exchange::Client.use_real_client
+    OpenStax::Exchange::use_real_client
   end
 
   it_behaves_like "exchange client api v1"
