+++ spec/lib/openstax/exchange/client/shared_examples_for_exchange_client_v1.rb
@@ -3,8 +3,8 @@ require 'spec_helper'
 RSpec.shared_examples "exchange client api v1" do
 
   before(:each) do
-    OpenStax::Exchange::Client.reset!
-    OpenStax::Exchange::Client.configure do |config|
+    OpenStax::Exchange.reset!
+    OpenStax::Exchange.configure do |config|
       config.platform_id     = DEFAULT_CLIENT_PLATFORM_ID
       config.platform_secret = DEFAULT_CLIENT_PLATFORM_SECRET
       config.server_base_url = DEFAULT_CLIENT_SERVER_BASE_URL
@@ -16,48 +16,48 @@ RSpec.shared_examples "exchange client api v1" do
   describe "#initialize" do
     context "success" do
       it "initializes the authentication token" do
-        client = OpenStax::Exchange::Client.send(:client)
+        client = OpenStax::Exchange.send(:client)
         expect(client.token).to_not be_nil
       end
     end
     context "invalid server" do
       it "raises an exception" do
-        OpenStax::Exchange::Client.configure do |config|
+        OpenStax::Exchange.configure do |config|
           config.server_base_url = 'http://this.is.a.fake.address'
         end
         expect {
-          OpenStax::Exchange::Client.send(:client)
-        }.to raise_error(OpenStax::Exchange::Client::ClientError)
+          OpenStax::Exchange.send(:client)
+        }.to raise_error(OpenStax::Exchange::ClientError)
       end
     end
     context "invalid port" do
       it "raises an exception" do
-        OpenStax::Exchange::Client.configure do |config|
+        OpenStax::Exchange.configure do |config|
           config.server_port = 9999
         end
         expect {
-          OpenStax::Exchange::Client.send(:client)
-        }.to raise_error(OpenStax::Exchange::Client::ClientError)
+          OpenStax::Exchange.send(:client)
+        }.to raise_error(OpenStax::Exchange::ClientError)
       end
     end
     context "invalid platform id" do
       it "raises an exception" do
-        OpenStax::Exchange::Client.configure do |config|
+        OpenStax::Exchange.configure do |config|
           config.platform_id = '999'
         end
         expect {
-          OpenStax::Exchange::Client.send(:client)
-        }.to raise_error(OpenStax::Exchange::Client::ClientError)
+          OpenStax::Exchange.send(:client)
+        }.to raise_error(OpenStax::Exchange::ClientError)
       end
     end
     context "invalid platform secret" do
       it "raises an exception" do
-        OpenStax::Exchange::Client.configure do |config|
+        OpenStax::Exchange.configure do |config|
           config.platform_secret = 'not_the_secret'
         end
         expect {
-          OpenStax::Exchange::Client.send(:client)
-        }.to raise_error(OpenStax::Exchange::Client::ClientError)
+          OpenStax::Exchange.send(:client)
+        }.to raise_error(OpenStax::Exchange::ClientError)
       end
     end
   end
@@ -65,12 +65,12 @@ RSpec.shared_examples "exchange client api v1" do
   describe "#create_identifier" do
     context "success" do
       it "creates and returns a new identifier" do
-        identifier = OpenStax::Exchange::Client.create_identifier
+        identifier = OpenStax::Exchange.create_identifier
         expect(identifier).to match(/^[a-fA-F0-9]+$/)
       end
       it "creates a distinct identifer per invokation" do
-        identifier1 = OpenStax::Exchange::Client.create_identifier
-        identifier2 = OpenStax::Exchange::Client.create_identifier
+        identifier1 = OpenStax::Exchange.create_identifier
+        identifier2 = OpenStax::Exchange.create_identifier
         expect(identifier1).to_not eq(identifier2)
       end
     end
@@ -79,7 +79,7 @@ RSpec.shared_examples "exchange client api v1" do
   describe "#create_multiple_choice" do
     context "success" do
       it "creates a multiple choice response associated with the given identifier" do
-        identifier = OpenStax::Exchange::Client.create_identifier
+        identifier = OpenStax::Exchange.create_identifier
 
         # must have the form of a "trusted resource"
         # (Exchange app/routines/find_or_create_resource_from_url.rb:35)
@@ -87,7 +87,7 @@ RSpec.shared_examples "exchange client api v1" do
         trial           = '1'
         answer_string   = 'answer_string'
 
-        response = OpenStax::Exchange::Client.create_multiple_choice(
+        response = OpenStax::Exchange.create_multiple_choice(
           identifier, resource_string, trial, answer_string)
 
         expect(response['identifier']).to eq(identifier)
@@ -96,8 +96,8 @@ RSpec.shared_examples "exchange client api v1" do
         expect(response['answer']).to eq(answer_string)
       end
       it "allows answers with distinct identifiers to be saved" do
-        identifier1 = OpenStax::Exchange::Client.create_identifier
-        identifier2 = OpenStax::Exchange::Client.create_identifier
+        identifier1 = OpenStax::Exchange.create_identifier
+        identifier2 = OpenStax::Exchange.create_identifier
 
         # must have the form of a "trusted resource"
         # (Exchange app/routines/find_or_create_resource_from_url.rb:35)
@@ -108,19 +108,19 @@ RSpec.shared_examples "exchange client api v1" do
         response = nil
 
         expect {
-          response = OpenStax::Exchange::Client.create_multiple_choice(
+          response = OpenStax::Exchange.create_multiple_choice(
             identifier1, resource_string, trial, answer_string)
         }.to_not raise_error
         expect(response['identifier']).to eq(identifier1)
 
         expect {
-          response = OpenStax::Exchange::Client.create_multiple_choice(
+          response = OpenStax::Exchange.create_multiple_choice(
             identifier2, resource_string, trial, answer_string)
         }.to_not raise_error
         expect(response['identifier']).to eq(identifier2)
       end
       it "allows answers with distinct resources to be saved" do
-        identifier = OpenStax::Exchange::Client.create_identifier
+        identifier = OpenStax::Exchange.create_identifier
 
         # must have the form of a "trusted resource"
         # (Exchange app/routines/find_or_create_resource_from_url.rb:35)
@@ -132,19 +132,19 @@ RSpec.shared_examples "exchange client api v1" do
         response = nil
 
         expect {
-          response = OpenStax::Exchange::Client.create_multiple_choice(
+          response = OpenStax::Exchange.create_multiple_choice(
             identifier, resource_string1, trial, answer_string)
         }.to_not raise_error
         expect(response['resource']).to eq(resource_string1)
 
         expect {
-          response = OpenStax::Exchange::Client.create_multiple_choice(
+          response = OpenStax::Exchange.create_multiple_choice(
             identifier, resource_string2, trial, answer_string)
         }.to_not raise_error
         expect(response['resource']).to eq(resource_string2)
       end
       it "allows answers with distinct trials to be saved" do
-        identifier = OpenStax::Exchange::Client.create_identifier
+        identifier = OpenStax::Exchange.create_identifier
 
         # must have the form of a "trusted resource"
         # (Exchange app/routines/find_or_create_resource_from_url.rb:35)
@@ -156,13 +156,13 @@ RSpec.shared_examples "exchange client api v1" do
         response = nil
 
         expect {
-          response = OpenStax::Exchange::Client.create_multiple_choice(
+          response = OpenStax::Exchange.create_multiple_choice(
             identifier, resource_string, trial1, answer_string)
         }.to_not raise_error
         expect(response['trial']).to eq(trial1)
 
         expect {
-          response = OpenStax::Exchange::Client.create_multiple_choice(
+          response = OpenStax::Exchange.create_multiple_choice(
             identifier, resource_string, trial2, answer_string)
         }.to_not raise_error
         expect(response['trial']).to eq(trial2)
@@ -170,7 +170,7 @@ RSpec.shared_examples "exchange client api v1" do
     end
     context "duplicate (identifer,resource,trial) triplet" do
       it "raises an exception" do
-        identifier = OpenStax::Exchange::Client.create_identifier
+        identifier = OpenStax::Exchange.create_identifier
 
         # must have the form of a "trusted resource"
         # (Exchange app/routines/find_or_create_resource_from_url.rb:35)
@@ -179,19 +179,19 @@ RSpec.shared_examples "exchange client api v1" do
         answer_string   = 'answer_string'
 
         expect {
-          response = OpenStax::Exchange::Client.create_multiple_choice(
+          response = OpenStax::Exchange.create_multiple_choice(
             identifier, resource_string, trial, answer_string)
         }.to_not raise_error
 
         expect {
-          response = OpenStax::Exchange::Client.create_multiple_choice(
+          response = OpenStax::Exchange.create_multiple_choice(
             identifier, resource_string, trial, answer_string)
-        }.to raise_error(OpenStax::Exchange::Client::ClientError)
+        }.to raise_error(OpenStax::Exchange::ClientError)
       end
     end
     context "invalid resource string" do
       it "raises an exception" do
-        identifier = OpenStax::Exchange::Client.create_identifier
+        identifier = OpenStax::Exchange.create_identifier
 
         # must have the form of a "trusted resource"
         # (Exchange app/routines/find_or_create_resource_from_url.rb:35)
@@ -200,9 +200,9 @@ RSpec.shared_examples "exchange client api v1" do
         answer_string   = 'answer_string'
 
         expect {
-          response = OpenStax::Exchange::Client.create_multiple_choice(
+          response = OpenStax::Exchange.create_multiple_choice(
             identifier, resource_string, trial, answer_string)
-        }.to raise_error(OpenStax::Exchange::Client::ClientError)
+        }.to raise_error(OpenStax::Exchange::ClientError)
       end
     end
   end
